Fix #2689: reject vector embeddings that underflow to zero in f32 (#2711)
* Fix #2689: reject vector embeddings that underflow to zero in f32

Subnormal-magnitude embedding components (|x| < ~1.2e-38) underflowed to
0.0f32 when serde deserialized the wire `Vec<f32>` fields, and the engine's
finiteness check ran post-cast, so the zero passed and a silently-corrupted
(all-zero) vector was stored with applied=true. The opposite extreme (inf) was
caught, so the two extremes were asymmetric.

Widen the upsert wire embedding fields to f64 (Command::VectorUpsert.vector,
BatchVectorEntry.vector, and the CLI parser) and add engine
VectorEmbedding::from_wire, which narrows f64 to the stored f32 and rejects a
non-zero magnitude that underflows to zero — with the same
invalid_argument.engine.vector_embedding as the overflow case. Stored format is
unchanged (still f32; bit-identical for representable values). Query vectors
keep f32 and are tracked separately (#2710).

Invariants verified: SCALE-003 (HOLDS), ARCH-003 (HOLDS) — ingest validation
only; stored representation and write path unchanged.
Tests added: from_wire_rejects_non_representable_values_and_keeps_genuine_zero, vector_upsert_rejects_a_subnormal_embedding_instead_of_storing_zeros


* Regenerate IDL docs for the vector.upsert embedding field

The vector-embedding field's doc comment changed with the f64-ingest fix;
regenerate the generated docs so the check-docs drift gate passes.


---------

+++ crates/executor/src/command.rs
@@ -661,8 +661,9 @@ pub enum Command {
         collection: String,
         /// Vector key.
         key: String,
-        /// Dense embedding.
-        vector: Vec<f32>,
+        /// Dense embedding. Accepted at wire (f64) precision and narrowed to the
+        /// stored f32; a value that underflows or overflows f32 is rejected.
+        vector: Vec<f64>,
         /// Optional metadata.
         #[serde(default, skip_serializing_if = "Option::is_none")]
         metadata: Option<Value>,
